Simplify GitConfig handling
Repository::open_from_env() finds the repo, respecting $GIT_DIR (which,
as I understand, is supposed to be always specified when we're invoked
via `git rebase --interactive`).

If we're invoked with $GIT_DIR unset (when running directly from the
command line?) the old code fell back to using the global config,
whereas the new version attempts to find the git repo from the current
directory, or exits with an error if not found, which I think is fine.

repo.config() returns a combined configuration object from the local
repo's config (.git/config) and the global/system configuration: it's
not documented very well, but true:
https://github.com/libgit2/libgit2/blame/814389d400b17cf5cce7862ca802d5a3eedba541/src/repository.c#L952