Core application rewrite
This change rewrites most of the core of the application from the ground up.

The most significant change is the handling of text overflow. Horizontally text
is truncated to fit the width of the terminal and vertical overflow is handled
with scrolling. Finally, when the terminal window gets too small a compact view
is rendered instead of the full normal view.

Previously the show commit pane would capture and parse stdout from an external
git command. Since this is slow and prone to error, this has been updated to
instead use libgit2 (using git2-rs) to retrieve information about the commit.

All functionality in the application is now handled using the internal state
machine. Initially certain functionality, such as the dialogs and shelling out
to an external editor was not using the state machine. This resulted in these
operations being inconsistent.