Improve performance of show commit
This change adds two caches that greatly speed up show commit. The first
is to cache the Unicode lengths so that they do not need to be
calculated for every render. The second is to ensure that the call to
libgit2 for retrieving the commit information is only made once and not
on every render loop. This should reduce the overall render time by
about ten fold.