Fix broken colours in Windows.
The cast from the pancurses chtyped to i16 did not work as expected on
Windows. This change instead uses the chtype directly. As well, Windows
only supports a total of 16 colour pairs and the tool requires over 16
total colour pairs. This is fixed by ensuring over 16 colour pairs instead
of 8 pairs.