Fix ANSI colour support
The version of curses that is used on Unix like systems does not
properly support truecolor. Because of this, existing colour codes
indexes were being overwritten to provide support for arbitrary custom
colours. This worked fine in the terminal emulators tested on Linux, as
the colours were reset back to their originals when the tool exited.
However, on MacOS, the colour changes would remain after the tool exited,
until the terminal was restarted.

To fix this, in terminals that do not support overwriting colours, a
provided colour triplet will be matched against the closest ANSI
indexed colour.