Fix scroll position resetting on resize
In several cases when the terminal was resized the scroll position was
reset back to the top. This was partially by design as a quick fix to
ensure that the top position did not go out of bounds. This change adds
a new function to the scroll position module that will ensure that the
scroll top value stays within the allowed range when the view is
resized.