Refactor and improve view line_segment
This change updates the line segment struct to return a segment partial
instead of directly drawing the line segment internally. This decouples
the line segment struct from the display module. The functionality for
drawing a line segment to the display has been migrated into the view.

Additionally tests have been added to the line segment struct to ensure
that the partial calculation for the line segment is performing as
expected.