Rewrite the view text rendering system
This updates the view text rendering system to render data based on a
data structure instead of having each module have to manually render the
required text. The new data structure is based on the data struct that
was created for the show_commit module, but has been expanded to provide
additional functionality and performance benefits. The new rendering
system has been implemented for the help, list and show_commit modules.

Additionally tests have been added to the view module to ensure that
this new functionality performs as expected.