Simplify comment char usage
The comment character handling was fairly awkward requiring the value to
be modified and passed around when it should not have been needed. This
change handles the "auto" value of the comment character within the
configuration and internalizes the usage of the comment character when
reading the rebase TODO file to the GitInteractive struct.