Convert window size values to usize
The window size was being provided as it's native type of an `i32`,
where it was often used with `usize` values. This change converts the
`i32` to a `usize` on read so that the conversion was not needed when it
was used. Since the window size should easily fit inside a `usize` on
any system that would be running git, this should be fine.