Add module test utilities and confirm abort tests
This adds a set of utility macros that assist in the testing of the
process module trait functions. A virtual curses interface is also
provided that allows the curses functionality to be mocked out. The
test utilities are not complete but provide enough functionality to
add tests for the rather simple confirm abort module.