Refactor process module input tests
The input module tests required knowledge of how Curses parses and
converts the raw inputs into the internal representation of inputs. This
resulted in fragile tests that would break if the key bindings change or
the usages of Curses changed. This updates those tests to take the
internal inputs instead of the Curses equivalent.