DRY up process module
The process module has always been fairly repetitive and adding a new
module required updates to several places in the file. This change
refactors the process module to use dynamic dispatch to call the
functions on the ProcessModule trait dynamically. This required the error,
help and window size error systems to be modified to implement the
ProcessModule. The show commit module has been refactored to use the
error module. And the state and modules have been extracted into their
own Modules struct.