Refactor and test external editor
The external editor module has not seen any attention in some time and
was showing some warts. This refactors the module to remove some dead
code and unneeded complexity. This allowed for the error process result
to be simplified, view to not be provided to the process function and
for two of the internal states to be removed/combined.

Tests have also been added to ensure that the module works as expected.