Use anyhow in the main application struct
The main application struct uses String errors, which make it difficult
to provide context on errors. This updates the struct to provide errors
with anyhow and fixes any references that used to depend on the String
errors. While working on the code, several poor usages of match have
been replaced with `map_or*`, `unwrap_or*` chains.