Greatly improve input handling and missing keys
This greatly improves the handling of keyboard input by ensuring that
all Pancurses inputs are properly accounted for. This resulted in
several missing keys being added. Also, provide a standard function for
handling the scrolling of view data and add scrolling support to the
error display. Also add End and Home support to the edit module.