Refactor the tests that use process utils
This greatly simplifies the process utils based tests to use a function
over a macro. The macro did not provide a decent stack-trace on error and
was also difficult to maintain. This change uses a basic function with a
callback that contains a test context that provides the flexibility
needed for future tests.