Use ViewLine::from trait instead of ViewLine::new
The From trait was meant to handle conversion from one type to another,
and the ViewLine struct did not take advantage of it. This removes the
ViewLine::new function and replaces the existing usages of the function
with a simplified version using the From trait. This greatly reduces the
complexity of the usages of ViewLine in most cases.