Remove process function from process module
The process function from the process module was only being used in the
external editor system. The usage in the external editor system was only
used as a faux on activate hook. Since the existing activate function
can instead be used to perform the same action, the process function can
be removed. This change removes the last usage of the process function
and then completely removes the usage of the process function from the
project.