Update licenses
Use cargo-about to generate a licenses file to list all third-party
library license. Provide a license flag to print the projects license as
well as a link to the generated licenses file.

The projects license file has been modified to use the short version of
the GPLv3 that links to the full body test. The licenses has not
changed.