Improve test render of padding
Previously the render of padding in tests would strip any default
padding from the end of lines, but fully print non-default padding. This
change keeps the stripping of default padding, but outputs a token
version of other types of padding. A test in process:help was also
updated to use the new type of padding.