Fix custom keybindings using modifiers
Almost all the modifiers with keybindings did not function as intended.
This updates the config parsing for keybindings to support all
modifiers. As a side effect, the modifiers can now be provided in any
order. Also, since the tests for this were fragile and required a config
file per test case, they have been refactored to generate a git
configuration as needed instead.