Make get_selected_line return an Option
This has been a long standing oversight in the todo_file module. It was
possible, generally due to an error, for the index of the todo list file
to be outside the allowed range, causing a crash. This changes the
method for getting the selected line to return an Option, to ensure that
the out of bounds case is handled.