Refactor CLI and main function
This replaces the CLI argument parsing with a lighter library and drops
the need for the heavy Clap library. Most of the main function code is
also moved into a new core module, and additional lighter tests are
added. A build.rs file has also been added to provide a more detailed
version info output.