Use Acquire-Release Ordering in atomic operations
The chosen Ordering was previously picked based on "getting things
working", and was not based on actual requirements. Technically Relaxed
is fine in all cases, since the poisoned value does not have any strong
data requirements. But since it is essentially acquire/release in almost
all cases, it makes sense to be explicit. The one use of sequential
consistent ordering in a test is confusing, and could always have been
relaxed or acquire/release.