Replace ProcessResult with Results
This replaces the ProcessResult type with a more robust Results struct.
This struct can define a set of result artifacts as a list, instead of
being limited to a single "result" set. This also allowed for the event
to be automatically added to the process result, instead of having to be
added for every event handler. Also this allows Process to include
additional response artifacts to be added dynamically based on the
processing of the existing artifacts.