Remove dynamic modules
Dynamically registrable modules were added a while back in an attempt
to make adding new modules easier. This has caused a number of issues
since its implementation, without much benefit. This removes the
ability to dynamically register a new module, instead providing a static
lookup on State to get a Module. This required significant test changes,
since being able to swap a module was key to a number of tests.