Fix race with ending thread and handling error
When an error occurred in the process thread, it was notifying the
runtime of the error condition before handling the error. This meant
this in some cases, the runtime would shutdown threads before the
process thread had handled the error. Since the error state is supposed
to be final, it is wrong for the application to attempt to process data
after reaching that state.