Remove Result returns from view state
The view thread state returned a Result for any state update that used
the internal state update channel to message on the update. While the
send call technically returns a Result that could be an error, it will
always be successfully since an unbounded channel cannot have a send
error.