Remove dependency between display and input crates
The input crate had an implicit dependency on the display crate, since
the read_event function had to be provided to the event provider from
the display crate. With this change, that dependency has been removed
and all event handling is now part of the input crate. The changes
include:

Moving the Tui::read_event function into the input crate. This new
function is not tied to Tui or any wrapper around Crossterm, and instead
is a standalone function that uses the "static" crossterm crate event
reading functions.

Adding a EventReaderFn trait with default implementations for all
functions that meet the requirements. This greatly simplifies the method
signature for the event provider.

Removal of the RawEvent export. This export was added as a quick fix to
allow some tests to generate the events from crossterm directly. This
was very useful for testing the application at the higher levels, where
the event provider could not be easily injected. Instead of using
RawEvent, a new test utility, create_event_reader, was added that will
map events into their internal crossterm implementation.