Use dynamic width in list action name render
Some upcoming changes will result in the action name no longer being
at a maximum of 6 characters. This change calculates the maximum action
width, based on the actions currently available in the list. This
required many of the test to be reworked, so where possible the new
EXCLUDE_STYLE option was used for render tests, to greatly reduce the
amount of noise in tests that did not need to verify style information.