Use shared reference to TodoFile in core crate
Previously, the TodoFile reference was passed through the Module
methods, but this resulted in a messy interface, where modules that did
not need to use the TodoFile reference ended up having an ignored
parameter. Also, if the TodoFile needed to be accessed outside of the
methods of the Module, then the modules got more complicated.

This change, adds a shared Arc<Mutex<TodoFile>> to all modules and
related processing code, and removes the method references from the
Module trait.