Normalize Shift modifier for character keycodes
Depending on the platform, reporting of the SHIFT modifier is not always
supported. This results in cases where keybindings that usually use the
SHIFT modifier (such as the ? character) not working as expected.

This change updates the keybindings, to remove and ignore any SHIFT
modifiers for any character keybinding, as well as convert the character
to the uppercase variant. To reflect this change, the event handler has
been changed to remove the SHIFT modifier for all characters, instead of
the old behavior of removing the SHIFT modifier only for uppercase ASCII
characters.