Fix sharing of ThreadStatus
Application was creating two instances of Runtime, the first instance
was only being used to create a ThreadStatus that was passed to Process.
The second instance was the real Runtime, that would handle all the
thread management and status tracking. The problem, was that Process had
an incorrect copy of ThreadStatus, so the status checks Process
performed were not completing. This resulted in things like the External
Editor functionality stalling, when it was waiting for thread status
updated.

This change creates a single instance of the ThreadStatus that is shared
across the Process and Runtime.