Add Operation::Load to History
Add to the todo_file History a load operation tracker, that will acts as
a sentinel value for the start of the undo history queue. This will
enable hooking into the start of the history queue. In order to expose
the additional information, the Operation is now returned from the
History undo/redo methods.