Fix flicker caused by action width change
When the longest action name changed, it would cause the action column
in the list display to resize, causing a flicker. While this was
intended behavior, it almost always caused a redraw, which was jarring.
This changes the rendering of the action to keep the minimum draw width
of the action at 6, to reduce the amount of flicker.