Properly support multiple visible lines in view
Previously, only one line could be marked as visible in the view data.
This resulted in the multi-select mode not correctly scrolling to the
start index in cases where lines are reordered.

This change adds support for multiple visible lines in the view data,
resulting in the "ensure visible" algorithm running over each visible
line.