Fix hidden line on non-body line addition
When a leading, trailing or header line was added to the view data, the
RenderSlice was not correctly updated to take into account which lines
should be visible. This resulted in some lines being hidden below the
view "window". This fixes the RenderSlice to recalculate the scroll
position when a leading, trailing or header line is changed.