Update release process
This change contains a major rewrite of the release process to generate
fewer overall releases, while adding support for additional targets. The
most significant change is releasing fully static binaries for Linux
targets, meaning there is no longer a need to make multiple binaries to
support different versions of glibc. macOS now has a universal binary
that will work on Apple silicon as well as x86_64 systems. On Windows a
new target of x86_64 GNU is now being published.

The other noteworthy change is the creation of archives for each
release, where each archive includes the README, license and other
related files, along with the binary.

On a project development point of view, the latest release and full
release now shares a pipeline. This ensures that the release pipeline is
being regularly actioned, as well as providing latest releases against
all supported targets.